+++ lib/src/query_results.dart
@@ -1,0 +1,5 @@
+class QueryResults{
+  final List<Map<String, dynamic>>? results;
+
+  QueryResults({this.results});
+}