+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0
+- Using now `QueryResults` as return object to the `sql` function
+
+## 0.3.2
+- Added shorter getter to user's ID (`SupabaseClient.uid` instead `SupabaseClient.auth.currentUser.uid`)
+
 ## 0.3.1
 - Fixed bug with `SELECT` statements like `..AND field = 'string'`